Improve, fix and add features

+++ src/system/config/aur.sh
@@ -26,11 +26,13 @@ function install_aur() {
         export AUR="paru"
     fi
 
-    exec_log "git clone ${git_url[$id]}" "Cloning ${aur_name[$id]}"
-    cd "${aur_name[$id]}" || return 1
-    exec_log "makepkg -si --noconfirm" "Installing ${AUR}"
-    cd .. || return 1
-    exec_log "rm -rf ${aur_name[$id]}" "Deleting directory ${aur_name[$id]}"
+    if ! pacman -Qi "${AUR}" &>/dev/null; then
+        exec_log "git clone ${git_url[$id]}" "Cloning ${aur_name[$id]}"
+        cd "${aur_name[$id]}" || return 1
+        exec_log "makepkg -si --noconfirm" "Installing ${AUR}"
+        cd .. || return 1
+        exec_log "rm -rf ${aur_name[$id]}" "Deleting directory ${aur_name[$id]}"
+    fi
 
     if [[ $choice == "yay" ]]; then
         exec "yay -Y --gendb" "Configuring ${AUR}"
@@ -40,6 +42,9 @@ function install_aur() {
         exec "paru --gendb" "Configuring ${AUR}"
         exec_log "sudo sed -i 's/#BottomUp/BottomUp/' /etc/paru.conf" "Enabling BottomUp option for paru"
         exec_log "sudo sed -i 's/#SudoLoop/SudoLoop/' /etc/paru.conf" "Enabling SudoLoop option for paru"
+        exec_log "sudo sed -i 's/#CombinedUpgrade/CombinedUpgrade/' /etc/paru.conf" "Enabling CombinedUpgrade option for paru"
+        exec_log "sudo sed -i 's/#UpgradeMenu/UpgradeMenu/' /etc/paru.conf" "Enabling UpgradeMenu option for paru"
+        exec_log "sudo sed -i 's/#NewsOnUpgrade/NewsOnUpgrade/' /etc/paru.conf" "Enabling NewsOnUpgrade option for paru"
         exec_log "sudo sh -c 'echo "SkipReview" >> /etc/paru.conf'" "Enabling SkipReview option for paru"
     fi
 }
